move update-api-policies to dedicated file
